app3_redis/requirements.txt
pyyaml==5.4.1
redis==4.3.4
Flask==2.1.3